NPC_t: use vbint_t for Direction
int8_t is possible as well if needed

+++ src/main/record.cpp
@@ -642,7 +642,7 @@ static void write_NPCs()
         fprintf(record_file, "NPC %d\r\n", i);
         fprintf(record_file, "Type %d\r\n", n.Type);
         fprintf(record_file, "Active %d\r\n", n.Active);
-        fprintf(record_file, "Dir %lf\r\n", n.Direction);
+        fprintf(record_file, "Dir %lf\r\n", (double)n.Direction);
         fprintf(record_file, "XYWH %lf %lf %lf %lf\r\n", n.Location.X, n.Location.Y, n.Location.Width, n.Location.Height);
         fprintf(record_file, "S %lf %lf %lf %lf %lf %lf %lf\r\n", n.Special, n.Special2, n.Special3, n.Special4, n.Special5, n.Special6, (double)n.Variant);
     }
